Rename `RcDropdown.d.ts` to `types.ts`
