+++ pkg/rancher-components/src/components/RcDropdown/types.ts
@@ -1,9 +1,13 @@
+import { Ref, ref } from 'vue';
+import type { RcButtonType } from '@components/RcButton';
+
 export type DropdownContext = {
   handleKeydown: () => void;
   showMenu: (show: boolean) => void;
   registerTrigger: (triggerRef: RcButtonType | null) => void;
   focusFirstElement: () => void;
   isMenuOpen: Ref<boolean>;
+  close: () => void;
 }
 
 export const defaultContext: DropdownContext = {
@@ -12,7 +16,7 @@ export const defaultContext: DropdownContext = {
   registerTrigger:   (_triggerRef: RcButtonType | null) => null,
   focusFirstElement: () => null,
   isMenuOpen:        ref(false),
-  close: () => null,
+  close:             () => null,
 };
 
 export type DropdownCollection = {
@@ -20,4 +24,3 @@ export type DropdownCollection = {
 };
 
 export const defaultCollection: DropdownCollection = { dropdownItems: ref([]) };
-
